Done lots of testing with various RNG seeds to see what evolves from given networks, starting with a disembodied simulation to see how it evolves and fix several bugs.
Got some interesting results, especially from seed 73, and ran it overnight to see how far it could get and saved it's resulting network.

I'm going to measure it's speed roughly in terms of firing rates, i.e. how fast a think_i can occur. For now it was doing about 15 iterations of 100 thinks per second, so that's about 1500Hz. Want this to be MhZ if possible, it could evolve much much quicker. Also going to upgrade to 64x64, it's way too small otherwise. But for now, speed.