Refactored entire system to use data structures optimized around most costly operations in its usage; now uses numpy arrays and linalg ops for propagation and most static ops (since changing np arrays is not fast).
Also found a bug in the RNG generation which resulted in them not being connected properly in the core! This might help things a lot.

Initial speed tests show this as around 35s/100k thinks, vs. old 64s/100k thinks. Exact improvement is 45.27% cutdown on time to think. hell yea.